implement and use Stratum:scaleToFit
