+++ client/lib/ArtifactStratum/render.js
@@ -32,4 +32,6 @@ module.exports = function () {
 
   // Re-compute bounding circle at each render.
   this.recomputeBoundingCircle()
+  // Scale so that cards just fit inside the rendering area.
+  this.scaleToFit()
 }
