+++ client/lib/Stratum/index.js
@@ -80,3 +80,4 @@ proto.getSubcontext = require('./getSubcontext')
 proto.getSupercontext = require('./getSupercontext')
 proto.recomputeBoundingCircle = require('./recomputeBoundingCircle')
 proto.render = require('./render')
+proto.scaleToFit = require('./scaleToFit')
